Added support for git submodules.
This PR is based in previous work by @jgrasser see: #1041, #1048 and #1469; it has been discussed also in other context as in #1479.
The code was base on those requests which are not of my authory, I recicled what @jgrasser made and make it compatible with the current code base.

It provides:
 * Support for git submodules
 * A bit of documentation on the feature
 * Unit test